+++ site/assets/backups/export-profile/site-rockfrontend/modules/RockDevTools/classes/JsArray.php
@@ -10,16 +10,25 @@ class JsArray extends FilenameArray
 {
   public function saveJS(string $to): void
   {
+    $js = '';
     if (str_ends_with($to, '.min.js')) {
       // minify content
-      $minifier = new JS();
-      foreach ($this as $file) $minifier->add($file);
+      foreach ($this as $file) {
+        $js .= ';'; // fix concatenating issues
+        if (str_ends_with(strtolower($file), '.min.js')) {
+          // add file as is
+          $js .= @wire()->files->fileGetContents($file);
+        } else {
+          $minifier = new JS();
+          $minifier->add($file);
+          $js .= $minifier->minify();
+        }
+      }
       $minifier->minify($to);
     } else {
       // merge content
-      $css = '';
-      foreach ($this as $file) $css .= @wire()->files->fileGetContents($file);
-      wire()->files->filePutContents($to, $css);
+      foreach ($this as $file) $js .= @wire()->files->fileGetContents($file);
     }
+    wire()->files->filePutContents($to, $js);
   }
 }